fixup! Add ChaCha related ciphers to default provider
