+++ test/poly1305_internal_test.c
@@ -14,7 +14,6 @@
 
 #include "testutil.h"
 #include "crypto/poly1305.h"
-#include "../crypto/poly1305/poly1305_local.h"
 #include "internal/nelem.h"
 
 typedef struct {
